Git init
Site is made by me to test JavaScript. Wanted to know what could be done with it and how easy that would be. Despite my expectations from the web it isn't really that horrible to use, which I now know. Also this was an idea that I had that seemed fun to make, so I made it